+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 1.11.0 /2018-08-03
+* Added a Simp::BeakerHelpers::WebServe module to allow for a local web server
+  to be spawned against a local directory for simple testing requiring external
+  web artifacts.
+
 ### 1.10.14 / 2018-08-01
 * Pinned `net-telnet` to `~> 0.1.1` for all releases due to dropping support
   for Ruby less than 2.3 in `0.2.X`. This should be removed once we drop